fix(palmtopd): first real Windows compiler errors
The Windows host had never been compiled until now (no Windows toolchain
was available in the environment that wrote it). Six errors from its first
real build, all fixed:

Three were one root cause -- missing windows-rs features, not wrong API
names. That crate gates individual *items* on the feature providing their
parameter/field types:
  - D3D11_TEXTURE2D_DESC and ID3D11Device::CreateTexture2D need
    Win32_Graphics_Dxgi_Common (the desc struct embeds DXGI_FORMAT)
  - Direct3D11CaptureFramePool::CreateFreeThreaded and
    Direct3D11CaptureFrame::Surface need Graphics_DirectX_Direct3D11
    (they take/return IDirect3DDevice / IDirect3DSurface)

MOUSEINPUT.mouseData is typed u32, but Win32 reads it back as a signed
delta for MOUSEEVENTF_WHEEL/HWHEEL -- scrolling up is negative. Cast at
the boundary; the parameter stays i32.

And one genuine surprise in *existing* Linux code: socket2 doesn't expose
TcpKeepalive::with_retries on Windows (or OpenBSD/Redox/Solaris), because
those platforms don't allow setting the retry count per socket. Now
cfg-gated. Not a behavioural gap -- keepalive time and interval are what
actually bound how long a vanished phone wedges a session, and Windows
applies its own fixed retry count on top. This is exactly the class of
thing that only surfaces by really compiling for the target, which is the
argument for the native Windows CI job that found it.

Encouragingly, the error spread across all three files means rustc
type-checked fully, so EnumDisplayMonitors, D3D11CreateDevice,
CreateForMonitor, FrameArrived and SetIsBorderRequired were all accepted
as written.

Linux verified unaffected: build, full test suite, and clippy all pass.

+++ crates/palmtopd/src/platform/windows/input.rs
@@ -125,7 +125,13 @@ fn send_mouse(dx: i32, dy: i32, mouse_data: i32, flags: windows::Win32::UI::Inpu
             mi: MOUSEINPUT {
                 dx,
                 dy,
-                mouseData: mouse_data,
+                // `mouseData` is typed u32, but for MOUSEEVENTF_WHEEL/HWHEEL
+                // Win32 reads it back as a *signed* delta -- scrolling up and
+                // scrolling left are negative. The cast is the two's-complement
+                // reinterpretation the API expects, not a lossy conversion, so
+                // the parameter stays i32 and is narrowed only here at the
+                // boundary.
+                mouseData: mouse_data as u32,
                 dwFlags: flags,
                 time: 0,
                 dwExtraInfo: extra,
